refactor: merge Sky and its loader to reduce unnecessary hierarchy
