+++ client/lib/Sky/findCurrentNode.js
