+++ client/lib/Sky/findCurrentStratum.js
